+++ src/app/app.component.ts
@@ -3,7 +3,6 @@ import { Nav, Platform, } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
 import { HomePage } from '../pages/home/home';
-import { Geolocation } from '@ionic-native/geolocation';
 
 @Component({
   templateUrl: 'app.html'
@@ -19,26 +18,12 @@ export class MyApp {
     public platform: Platform,
     public statusBar: StatusBar,
     public splashScreen: SplashScreen,
-    private geolocation: Geolocation,
     ) {
     this.initializeApp();
 
     this.pages = [
       { title: 'Home', component: HomePage },
         ];
-
-        this.geolocation.getCurrentPosition().then((resp) => {
-
-            }).catch((error) => {
-          console.log('Error getting location', error);
-        });
-
-        let watch = this.geolocation.watchPosition();
-        watch.subscribe((data) => {
-         // data can be a set of coordinates, or an error (if an error occurred).
-         // data.coords.latitude
-         // data.coords.longitude
-       });
   }
 
   initializeApp() {
